tco_app/requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24.3
plotly>=5.18.0
openpyxl>=3.1.2
Pillow>=10.0.0
pytest-cov>=4.1.0
ruff>=0.4
black>=24.4
numba>=0.58.0
